brochureScript/requirements.txt
dnspython==1.16.0
img2pdf==0.3.3
mysql-connector-python==8.0.19
pdf2image==1.12.1
Pillow==7.0.0
protobuf==3.6.1
pyodbc==4.0.30
PyPDF2==1.26.0
six==1.14.0
